Ajuste transferência de argumentos entre telas

+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/stable/Colab.kt
@@ -1,8 +1,6 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.stable
 
-import kotlinx.serialization.Serializable
-
 data class Colab (
-    val matricColab: Long,
+    val matricColab: Int,
     val nomeColab: String,
 )
