+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/stable/Equip.kt
@@ -1,6 +1,6 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.stable
 
 data class Equip (
-    val idEquip: Long,
+    val idEquip: Int,
     val nroEquip: Long,
 )
