+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/stable/Local.kt
@@ -1,6 +1,6 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.stable
 
 data class Local (
-    val idLocal: Long,
+    val idLocal: Int,
     val descrLocal: String,
 )
