+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/stable/Terceiro.kt
@@ -1,8 +1,8 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.stable
 
 data class Terceiro(
-    val idTerceiro: Long,
-    val idBDTerceiro: Long,
+    val idTerceiro: Int,
+    val idBDTerceiro: Int,
     val cpfTerceiro: String,
     val nomeTerceiro: String,
     val empresaTerceiro: String,
