+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/stable/Visitante.kt
@@ -1,7 +1,7 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.stable
 
 data class Visitante(
-    val idVisitante: Long,
+    val idVisitante: Int,
     val cpfVisitante: String,
     val nomeVisitante: String,
     val empresaVisitante: String,
