+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/variable/Config.kt
@@ -7,11 +7,11 @@ import br.com.usinasantafe.pcpcomp.utils.StatusSend
 data class Config(
     var number: Long? = null,
     var password: String? = null,
-    var idBD: Long? = null,
+    var idBD: Int? = null,
     var version: String? = null,
     var flagUpdate: FlagUpdate = FlagUpdate.OUTDATED,
-    var matricVigia: Long? = null,
-    var idLocal: Long? = null,
+    var matricVigia: Int? = null,
+    var idLocal: Int? = null,
     var statusSend: StatusSend = StatusSend.SENT,
     var statusData: StatusData = StatusData.CLOSE,
 )
