+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/entities/variable/MovEquipProprioSeg.kt
@@ -1,7 +1,7 @@
 package br.com.usinasantafe.pcpcomp.domain.entities.variable
 
 data class MovEquipProprioSeg(
-    var idMovEquipProprioSeg: Long? = null,
-    var idMovEquipProprio: Long? = null,
-    var idEquipMovEquipProprioSeg: Long? = null,
+    var idMovEquipProprioSeg: Int? = null,
+    var idMovEquipProprio: Int? = null,
+    var idEquipMovEquipProprioSeg: Int? = null,
 )
