+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/repositories/stable/LocalRepository.kt
@@ -7,6 +7,6 @@ interface LocalRepository {
     suspend fun addAll(list: List<Local>): Result<Boolean>
     suspend fun deleteAll(): Result<Boolean>
     suspend fun getAll(): Result<List<Local>>
-    suspend fun getDescr(id: Long): Result<String>
+    suspend fun getDescr(id: Int): Result<String>
     suspend fun recoverAll(token: String): Result<List<Local>>
 }
