+++ app/src/main/java/br/com/usinasantafe/pcpcomp/domain/usecases/config/SetIdLocalConfig.kt
@@ -4,14 +4,14 @@ import br.com.usinasantafe.pcpcomp.domain.errors.UsecaseException
 import br.com.usinasantafe.pcpcomp.domain.repositories.variable.ConfigRepository
 
 interface SetIdLocalConfig {
-    suspend operator fun invoke(idLocal: Long): Result<Boolean>
+    suspend operator fun invoke(idLocal: Int): Result<Boolean>
 }
 
 class SetIdLocalConfigImpl(
     private val configRepository: ConfigRepository,
 ): SetIdLocalConfig {
 
-    override suspend fun invoke(idLocal: Long): Result<Boolean> {
+    override suspend fun invoke(idLocal: Int): Result<Boolean> {
         try {
             val resultConfig = configRepository.getConfig()
             if(resultConfig.isFailure)
