+++ app/src/main/java/br/com/usinasantafe/pcpcomp/external/room/dao/stable/LocalDao.kt
@@ -19,6 +19,6 @@ interface LocalDao {
     suspend fun getAll(): List<LocalRoomModel>
 
     @Query("SELECT descrLocal FROM $TB_LOCAL WHERE idLocal = :id")
-    suspend fun getDescr(id: Long): String
+    suspend fun getDescr(id: Int): String
 
 }
