+++ app/src/main/java/br/com/usinasantafe/pcpcomp/external/room/dao/variable/MovEquipVisitTercDao.kt
@@ -33,9 +33,9 @@ interface MovEquipVisitTercDao {
     suspend fun listStatusForeigner(statusForeigner: StatusForeigner): List<MovEquipVisitTercRoomModel>
 
     @Query("SELECT MAX(idMovEquipVisitTerc) FROM $TB_MOV_EQUIP_VISIT_TERC WHERE statusMovEquipVisitTerc = :status")
-    suspend fun lastIdStatus(status: StatusData): Long
+    suspend fun lastIdStatus(status: StatusData): Int
 
     @Query("SELECT * FROM $TB_MOV_EQUIP_VISIT_TERC WHERE idMovEquipVisitTerc = :idMov")
-    suspend fun getId(idMov: Long): MovEquipVisitTercRoomModel
+    suspend fun getId(idMov: Int): MovEquipVisitTercRoomModel
 
 }
