+++ app/src/main/java/br/com/usinasantafe/pcpcomp/infra/datasource/room/stable/ColabRoomDatasource.kt
@@ -4,7 +4,7 @@ import br.com.usinasantafe.pcpcomp.infra.models.room.stable.ColabRoomModel
 
 interface ColabRoomDatasource {
     suspend fun addAll(list: List<ColabRoomModel>): Result<Boolean>
-    suspend fun checkMatric(matric: Long): Result<Boolean>
-    suspend fun getNome(matric: Long): Result<String>
+    suspend fun checkMatric(matric: Int): Result<Boolean>
+    suspend fun getNome(matric: Int): Result<String>
     suspend fun deleteAll(): Result<Boolean>
 }
