+++ app/src/main/java/br/com/usinasantafe/pcpcomp/infra/models/room/stable/EquipRoomModel.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 @Entity(tableName = TB_EQUIP)
 data class EquipRoomModel (
     @PrimaryKey
-    val idEquip: Long,
+    val idEquip: Int,
     val nroEquip: Long,
 )
