+++ app/src/main/java/br/com/usinasantafe/pcpcomp/infra/models/room/stable/TerceiroRoomModel.kt
@@ -9,8 +9,8 @@ import kotlinx.serialization.Serializable
 @Entity(tableName = TB_TERCEIRO)
 data class TerceiroRoomModel (
     @PrimaryKey(autoGenerate = true)
-    val idTerceiro: Long,
-    val idBDTerceiro: Long,
+    val idTerceiro: Int,
+    val idBDTerceiro: Int,
     val cpfTerceiro: String,
     val nomeTerceiro: String,
     val empresaTerceiro: String,
