+++ app/src/main/java/br/com/usinasantafe/pcpcomp/infra/models/sharedpreferences/MovEquipProprioSharedPreferencesModel.kt
@@ -8,10 +8,10 @@ import java.util.Date
 data class MovEquipProprioSharedPreferencesModel(
     var dthrMovEquipProprio: Date = Date(),
     var tipoMovEquipProprio: TypeMov,
-    var idEquipMovEquipProprio: Long? = null,
-    var nroMatricColabMovEquipProprio: Long? = null,
+    var idEquipMovEquipProprio: Int? = null,
+    var nroMatricColabMovEquipProprio: Int? = null,
     var destinoMovEquipProprio: String? = null,
-    var nroNotaFiscalMovEquipProprio: Long? = null,
+    var nroNotaFiscalMovEquipProprio: Int? = null,
     var observMovEquipProprio: String? = null,
 )
