+++ app/src/main/java/br/com/usinasantafe/pcpcomp/infra/models/webservice/ConfigWebServiceModel.kt
@@ -9,7 +9,7 @@ data class ConfigWebServiceModelOutput(
 )
 
 data class ConfigWebServiceModelInput(
-    var idBD: Long,
+    var idBD: Int,
 )
