+++ app/src/main/java/br/com/usinasantafe/pcpcomp/utils/Token.kt
@@ -7,7 +7,7 @@ import java.security.MessageDigest
 import java.util.Locale
 
 
-fun token(number: Long, version: String, idBD: Long): String {
+fun token(number: Long, version: String, idBD: Int): String {
     var token = "PCP-${version}-$number-$idBD"
     val messageDigest = MessageDigest.getInstance("MD5")
     messageDigest.update(token.toByteArray(), 0, token.length)
